[DEV/MAIN-IMP] Enhanced proxy injection handling with error redaction and validation improvements

+++ src/api/proxies/proxies.controller.ts
@@ -138,7 +138,7 @@ export const proxiesController = createElysia({ prefix: "/proxies" })
     },
     {
       body: z.object({
-        id: z.coerce.number().positive(),
+        id: z.coerce.number().positive().int(),
       }),
     },
   );
